# src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Pactum Flow API"
basePath: "/api/flow/v1"
consumes:
  - "application/json"
  - "text/html"
produces:
  - "application/json"
  - "text/html"
  - "text/plain"
paths:
...
  /analyses/{id}:
    get:
      tags:
        - "Analyses"
      description: "get analysis with id"
      operationId: "getAnalysisById"
      parameters:
        - in: "path"
          name: "id"
          description: "analysis id to fetch"
          required: true
          type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Analysis"
    delete:
      tags:
        - "Analyses"
      description: "delete an analysis"
      operationId: "deleteAnalysis"
      parameters:
        - in: "path"
          name: "id"
          description: "analysis id to delete"
          required: true
          type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/HealthResponse"
    x-swagger-router-controller: "analysis.controller"
...
definitions:
...
  Analysis:
    type: "object"
    properties:
      projectId:
        type: "string"
      branch:
        type: "string"
      version:
        type: "string"
    required:
     - "projectId"
     - "branch"
     - "version"
...
  HealthResponse:
    description: "Health Response"
    type: "object"
    properties:
      message:
        type: "string"
    example:
      message: "OK"